+++ functions/package.json
@@ -19,7 +19,7 @@
     "dotenv": "^16.3.1",
     "express-multipart-file-parser": "^0.1.2",
     "firebase-admin": "^11.9.0",
-    "firebase-functions": "^4.2.0",
+    "firebase-functions": "^4.3.0",
     "fs": "^0.0.1-security",
     "graphql": "^16.7.1",
     "graphql-request": "^6.1.0",
