+++ functions/src/api/index.js
@@ -5,7 +5,6 @@ const accountAPI = require("./account");
 const filesAPI = require("./files");
 const threadRipper = require("./thread-ripper");
 const proxy = require("./proxy");
-const redpacketreveal = require("./redpacketreveal");
 
 // Define the routes for the /api/ endpoint
 router.get("/", (req, res) => res.send({ status: "nfttoolkit at ur service!" }));
@@ -21,8 +20,6 @@ router.use("/thread-ripper", threadRipper);
 
 // Proxy for cors requests from client
 router.use("/proxy", proxy);
-
-router.use('/redpacketreveal', redpacketreveal);
 
 // Export the router
 module.exports = router;
