+++ functions/src/index.js
@@ -1,7 +1,8 @@
 const functions = require("firebase-functions");
 const express = require("express");
 const cors = require("cors");
-const fileParser = require('express-multipart-file-parser')
+const fileParser = require("express-multipart-file-parser");
+const redpacketreveal = require("./utils/redpacketreveal");
 const { authenticate: authenticateAndPersistApiKey } = require("./utils/loopring/api");
 require("dotenv").config();
 
@@ -10,8 +11,6 @@ const app = express();
 
 // Enable CORS for all routes
 app.use(cors());
-
-app.use(fileParser)
 
 // Import your API routes
 const api = require("./api");
@@ -31,4 +30,14 @@ app.use("/", api);
 
 // Export the Express app as a Firebase Function
 exports.api = functions.https.onRequest(app);
+
+
+// Create an express app for the API
+const imageProcessor = express();
+// Enable CORS for all routes
+imageProcessor.use(cors());
+imageProcessor.use(fileParser);
+imageProcessor.use("/", redpacketreveal);
+
+exports.redpacketreveal = functions.https.onRequest(imageProcessor);
