+++ functions/src/utils/redpacketreveal/overlay.webp
